# docs/openapi-v1.yaml
openapi: 3.0.3
info:
  title: WrkPlan API
  version: 1.0.0
  description: API contract for Laravel-to-.NET migration.
servers:
  - url: /api/v1
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: wrkplan_session
  schemas:
...
    LoginRequest:
      type: object
      required: [password]
      properties:
        login:
          type: string
          example: john.smith
        email:
          type: string
          example: john@acme.com
        password:
          type: string
        corp_id:
          type: string
          example: ACME-IND
        remember:
          type: boolean
    TenantCreateRequest:
      type: object
      required: [company_name, corp_id, contact_name, contact_email, user_email, username]
      properties:
        company_name: { type: string }
        corp_id: { type: string }
        contact_name: { type: string }
        contact_email: { type: string }
        user_email: { type: string }
        username: { type: string }
        contact_password: { type: string }
        contact_phone: { type: string }
        address: { type: string }
        city: { type: string }
        country: { type: string }
        timezone: { type: string }
        status: { type: string, enum: [active, inactive, trial] }
        plan_id: { type: integer }
        billing_cycle: { type: string, enum: [monthly, quarterly, annual] }
        custom_rate: { type: number, format: double }
        currency: { type: string }
    SubscriptionAssignRequest:
      type: object
      required: [plan_id, billing_cycle]
      properties:
        plan_id: { type: integer }
        billing_cycle: { type: string, enum: [monthly, quarterly, annual] }
        start_date: { type: string, format: date }
        custom_rate: { type: number, format: double }
        currency: { type: string, example: USD }
        notes: { type: string }
    PlanWriteRequest:
      type: object
      required: [name, monthly_price, quarterly_price, annual_price]
      properties:
        name: { type: string }
        description: { type: string }
        monthly_price: { type: number, format: double }
        quarterly_price: { type: number, format: double }
        annual_price: { type: number, format: double }
        max_users: { type: integer }
        sort_order: { type: integer }
        is_active: { type: boolean }
        features:
          type: array
          items: { type: string }
...
security:
  - sessionAuth: []
paths:
  /auth/login:
    post:
      summary: Login with username/email and optional corp_id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login success
...
  /admin/tenants:
    get:
      summary: Paginated tenant list
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: per_page
          schema: { type: integer }
      responses:
        '200': { description: Success }
    post:
      summary: Create tenant and optional initial subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TenantCreateRequest'
      responses:
        '201': { description: Created }
...
  /admin/tenants/{tenant}/toggle-status:
    post:
      summary: Toggle tenant active/inactive status
      parameters:
        - in: path
          name: tenant
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Updated }
  /admin/tenants/{tenant}/assign-subscription:
    post:
      summary: Assign subscription with optional custom rate
      parameters:
        - in: path
          name: tenant
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriptionAssignRequest'
      responses:
        '201': { description: Created }
...
  /admin/plans:
    get:
      summary: List plans
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: is_active
          schema: { type: boolean }
        - in: query
          name: per_page
          schema: { type: integer }
      responses:
        '200': { description: Success }
    post:
      summary: Create plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanWriteRequest'
      responses:
        '201': { description: Created }
...
  /admin/plans/{plan}/toggle-status:
    post:
      summary: Toggle plan status
      parameters:
        - in: path
          name: plan
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Updated }
...
  /admin/invoices/{invoice}/payment:
    post:
      summary: Record invoice payment
      parameters:
        - in: path
          name: invoice
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Recorded }
  /admin/invoices/{invoice}/download-pdf:
    get:
      summary: Download invoice PDF
      parameters:
        - in: path
          name: invoice
          required: true
          schema: { type: integer }
      responses:
        '200': { description: File }
...
  /admin/audit:
    get:
      summary: Paginated audit logs
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: module
          schema: { type: string }
        - in: query
          name: date_from
          schema: { type: string, format: date }
        - in: query
          name: date_to
          schema: { type: string, format: date }
        - in: query
          name: per_page
          schema: { type: integer }
      responses:
        '200': { description: Success }
...